requirements.txt
opencv-python==4.7.0.72
numpy==1.24.3
pycoral==0.2.0
Pillow==9.5.0
tqdm==4.65.0
matplotlib==3.7.1
plyer==2.1.0  # For desktop notifications 